refactor: divided main.py into multiple submodules
